# docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "API para gerenciamento de agendamentos de envio de mensagens em lote para WhatsApp.\nEste backend apenas armazena e gerencia agendamentos via API REST — não envia mensagens.",
        "title": "WPP Scheduler API",
        "contact": {
            "name": "Suporte WPP Scheduler",
            "url": "https://github.com/ropehapi/kaizen-wpp-scheduler-backend"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
        "/api/v1/schedules": {
            "get": {
                "description": "Retorna agendamentos com suporte a filtros por status e paginação.\nValores padrão: page=1, limit=10. Limite máximo: 100.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Agendamentos"
                ],
                "summary": "Lista agendamentos",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Filtro por status (scheduled, sent, canceled)",
                        "name": "status",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Número da página (padrão: 1)",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Itens por página (padrão: 10, máx: 100)",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_ropehapi_kaizen-wpp-scheduler-backend_pkg_response.PaginatedResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/github_com_ropehapi_kaizen-wpp-scheduler-backend_internal_domain.Schedule"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_ropehapi_kaizen-wpp-scheduler-backend_pkg_response.APIResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_ropehapi_kaizen-wpp-scheduler-backend_pkg_response.APIResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "Cria um agendamento com mensagem, contatos, tipo e data programada.\nSe o tipo for \"recurring\", o campo \"frequency\" é obrigatório (daily, weekly, monthly).",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Agendamentos"
                ],
                "summary": "Cria um novo agendamento",
                "parameters": [
                    {
                        "description": "Dados do agendamento",
                        "name": "schedule",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_ropehapi_kaizen-wpp-scheduler-backend_internal_domain.CreateScheduleRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_ropehapi_kaizen-wpp-scheduler-backend_pkg_response.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/github_com_ropehapi_kaizen-wpp-scheduler-backend_internal_domain.Schedule"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_ropehapi_kaizen-wpp-scheduler-backend_pkg_response.APIResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_ropehapi_kaizen-wpp-scheduler-backend_pkg_response.APIResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/schedules/{id}": {
            "get": {
                "description": "Retorna um agendamento específico pelo UUID, incluindo seus contatos associados.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Agendamentos"
                ],
                "summary": "Busca agendamento por ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "UUID do agendamento",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_ropehapi_kaizen-wpp-scheduler-backend_pkg_response.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/github_com_ropehapi_kaizen-wpp-scheduler-backend_internal_domain.Schedule"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_ropehapi_kaizen-wpp-scheduler-backend_pkg_response.APIResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/github_com_ropehapi_kaizen-wpp-scheduler-backend_pkg_response.APIResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_ropehapi_kaizen-wpp-scheduler-backend_pkg_response.APIResponse"
                        }
                    }
                }
            },
            "put": {
                "description": "Atualiza campos de um agendamento existente. Todos os campos são opcionais.\nNão é possível atualizar agendamentos com status \"sent\".",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Agendamentos"
                ],
                "summary": "Atualiza um agendamento",
                "parameters": [
                    {
                        "type": "string",
                        "description": "UUID do agendamento",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Dados para atualização (campos opcionais)",
                        "name": "schedule",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_ropehapi_kaizen-wpp-scheduler-backend_internal_domain.UpdateScheduleRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_ropehapi_kaizen-wpp-scheduler-backend_pkg_response.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/github_com_ropehapi_kaizen-wpp-scheduler-backend_internal_domain.Schedule"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_ropehapi_kaizen-wpp-scheduler-backend_pkg_response.APIResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/github_com_ropehapi_kaizen-wpp-scheduler-backend_pkg_response.APIResponse"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/github_com_ropehapi_kaizen-wpp-scheduler-backend_pkg_response.APIResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_ropehapi_kaizen-wpp-scheduler-backend_pkg_response.APIResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/schedules/{id}/cancel": {
            "patch": {
                "description": "Altera o status de um agendamento para \"canceled\".\nNão é possível cancelar agendamentos já enviados ou já cancelados.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Agendamentos"
                ],
                "summary": "Cancela um agendamento",
                "parameters": [
                    {
                        "type": "string",
                        "description": "UUID do agendamento",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_ropehapi_kaizen-wpp-scheduler-backend_pkg_response.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/github_com_ropehapi_kaizen-wpp-scheduler-backend_internal_domain.Schedule"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_ropehapi_kaizen-wpp-scheduler-backend_pkg_response.APIResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/github_com_ropehapi_kaizen-wpp-scheduler-backend_pkg_response.APIResponse"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/github_com_ropehapi_kaizen-wpp-scheduler-backend_pkg_response.APIResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_ropehapi_kaizen-wpp-scheduler-backend_pkg_response.APIResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "github_com_ropehapi_kaizen-wpp-scheduler-backend_internal_domain.Contact": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "phone": {
                    "type": "string"
                },
                "scheduleId": {
                    "type": "string"
                }
            }
        },
        "github_com_ropehapi_kaizen-wpp-scheduler-backend_internal_domain.CreateContactRequest": {
            "type": "object",
            "required": [
                "name",
                "phone"
            ],
            "properties": {
                "name": {
                    "type": "string"
                },
                "phone": {
                    "type": "string"
                }
            }
        },
        "github_com_ropehapi_kaizen-wpp-scheduler-backend_internal_domain.CreateScheduleRequest": {
            "type": "object",
            "required": [
                "contacts",
                "message",
                "scheduledAt",
                "type"
            ],
            "properties": {
                "contacts": {
                    "type": "array",
                    "minItems": 1,
                    "items": {
                        "$ref": "#/definitions/github_com_ropehapi_kaizen-wpp-scheduler-backend_internal_domain.CreateContactRequest"
                    }
                },
                "frequency": {
                    "$ref": "#/definitions/github_com_ropehapi_kaizen-wpp-scheduler-backend_internal_domain.Frequency"
                },
                "message": {
                    "type": "string"
                },
                "scheduledAt": {
                    "type": "string"
                },
                "type": {
                    "enum": [
                        "once",
                        "recurring"
                    ],
                    "allOf": [
                        {
                            "$ref": "#/definitions/github_com_ropehapi_kaizen-wpp-scheduler-backend_internal_domain.ScheduleType"
                        }
                    ]
                }
            }
        },
        "github_com_ropehapi_kaizen-wpp-scheduler-backend_internal_domain.Frequency": {
            "type": "string",
            "enum": [
                "daily",
                "weekly",
                "monthly"
            ],
            "x-enum-varnames": [
                "FrequencyDaily",
                "FrequencyWeekly",
                "FrequencyMonthly"
            ]
        },
        "github_com_ropehapi_kaizen-wpp-scheduler-backend_internal_domain.Schedule": {
            "type": "object",
            "properties": {
                "contacts": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_ropehapi_kaizen-wpp-scheduler-backend_internal_domain.Contact"
                    }
                },
                "createdAt": {
                    "type": "string"
                },
                "frequency": {
                    "$ref": "#/definitions/github_com_ropehapi_kaizen-wpp-scheduler-backend_internal_domain.Frequency"
                },
                "id": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                },
                "scheduledAt": {
                    "type": "string"
                },
                "status": {
                    "$ref": "#/definitions/github_com_ropehapi_kaizen-wpp-scheduler-backend_internal_domain.ScheduleStatus"
                },
                "type": {
                    "$ref": "#/definitions/github_com_ropehapi_kaizen-wpp-scheduler-backend_internal_domain.ScheduleType"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
        "github_com_ropehapi_kaizen-wpp-scheduler-backend_internal_domain.ScheduleStatus": {
            "type": "string",
            "enum": [
                "scheduled",
                "sent",
                "canceled"
            ],
            "x-enum-varnames": [
                "ScheduleStatusScheduled",
                "ScheduleStatusSent",
                "ScheduleStatusCanceled"
            ]
        },
        "github_com_ropehapi_kaizen-wpp-scheduler-backend_internal_domain.ScheduleType": {
            "type": "string",
            "enum": [
                "once",
                "recurring"
            ],
            "x-enum-varnames": [
                "ScheduleTypeOnce",
                "ScheduleTypeRecurring"
            ]
        },
        "github_com_ropehapi_kaizen-wpp-scheduler-backend_internal_domain.UpdateScheduleRequest": {
            "type": "object",
            "properties": {
                "contacts": {
                    "type": "array",
                    "minItems": 1,
                    "items": {
                        "$ref": "#/definitions/github_com_ropehapi_kaizen-wpp-scheduler-backend_internal_domain.CreateContactRequest"
                    }
                },
                "frequency": {
                    "$ref": "#/definitions/github_com_ropehapi_kaizen-wpp-scheduler-backend_internal_domain.Frequency"
                },
                "message": {
                    "type": "string"
                },
                "scheduledAt": {
                    "type": "string"
                },
                "type": {
                    "enum": [
                        "once",
                        "recurring"
                    ],
                    "allOf": [
                        {
                            "$ref": "#/definitions/github_com_ropehapi_kaizen-wpp-scheduler-backend_internal_domain.ScheduleType"
                        }
                    ]
                }
            }
        },
        "github_com_ropehapi_kaizen-wpp-scheduler-backend_pkg_response.APIResponse": {
            "type": "object",
            "properties": {
                "data": {},
                "error": {
                    "type": "string"
                }
            }
        },
        "github_com_ropehapi_kaizen-wpp-scheduler-backend_pkg_response.PaginatedResponse": {
            "type": "object",
            "properties": {
                "data": {},
                "error": {
                    "type": "string"
                },
                "pagination": {
                    "$ref": "#/definitions/github_com_ropehapi_kaizen-wpp-scheduler-backend_pkg_response.PaginationInfo"
                }
            }
        },
        "github_com_ropehapi_kaizen-wpp-scheduler-backend_pkg_response.PaginationInfo": {
            "type": "object",
            "properties": {
                "limit": {
                    "type": "integer"
                },
                "page": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                },
                "totalPages": {
                    "type": "integer"
                }
            }
        }
    }
}